More refactoring, added abstract treeNode class
